test(admin): split widget visibility fixtures into separate files
PHPCS PSR1.Classes.ClassDeclaration rejects multiple classes per
file. Move each fixture widget into its own file under
tests/Filament/Admin/Traits/Fixtures/ with the matching namespace.
